+++ packages/typescript-eslint/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 8.18.2 (2024-12-23)
+
+### 🩹 Fixes
+
+- **typescript-eslint:** export types so that declarations can be named for dts files ([#10513](https://github.com/typescript-eslint/typescript-eslint/pull/10513))
+
+### ❤️ Thank You
+
+- Brad Zacher @bradzacher
+
+You can read about our [versioning strategy](https://main--typescript-eslint.netlify.app/users/versioning) and [releases](https://main--typescript-eslint.netlify.app/users/releases) on our website.
+
 ## 8.18.1 (2024-12-16)
 
 This was a version bump only for typescript-eslint to align it with other projects, there were no code changes.